Migrate PropertyEditModel to a Model-View-Controller architecture.
