+++ src/main/java/decodes/dbeditor/EquipmentEditPanel.java
@@ -63,8 +63,7 @@ public class EquipmentEditPanel extends DbEditorTab
     public EquipmentEditPanel()
 	{
 		Properties props = new Properties();
-		propertiesEditPanel =
-			new PropertiesEditPanel(props);
+		propertiesEditPanel = PropertiesEditPanel.from(props);
         try {
             jbInit();
         }
@@ -100,7 +99,7 @@ public class EquipmentEditPanel extends DbEditorTab
 			enableFields();
 			theObject = origObject.copy();
 			setTopObject(origObject);
-			propertiesEditPanel.setProperties(theObject.properties);
+			propertiesEditPanel.getModel().setProperties(theObject.properties);
 			fillFields();
 		}
 	}
@@ -114,7 +113,7 @@ public class EquipmentEditPanel extends DbEditorTab
 	{
 		enableFields();
 		theObject.copyFrom(imported);
-		propertiesEditPanel.setProperties(theObject.properties);
+		propertiesEditPanel.getModel().setProperties(theObject.properties);
 		fillFields();
 	}
 
@@ -183,7 +182,7 @@ public class EquipmentEditPanel extends DbEditorTab
 		if (TextUtil.isAllWhitespace(theObject.model))
 			theObject.model = null;
 		theObject.equipmentType = equipmentTypeSelector.getSelection();
-		propertiesEditPanel.saveChanges();
+		propertiesEditPanel.getModel().saveChanges();
 		return theObject;
 	}
